Remove TODO's
I'm happy with the current behaviour. I expect enable_linux_perf to be set globally when running and even if it's part of the benchmark definition, should be easy to override.
If `enable_linux_perf` is set correctly there should be no reason why an error might occur, so reasonable to throw.

I though the return_val_2 stuff might be unecessary as we don't do it normally but we do, it's just that we actually return the return_val where we don't here which necessitates the rand()